refactor(commands): extract flags tests to sibling file
